revised tbp example to current genesez development

+++ examples/trunk/org.genesez.example.java.tbp/src-gen/core/Duration.java
@@ -4,39 +4,33 @@ package core;
  *	Do not place import/include statements above this comment, just below. 
  * 	@FILE-ID : (_16_0_b6f02e1_1231923224937_24486_549) 
  */
-
 import org.genesez.platform.java.umlsupport.associations.*;
 
 /**
  * Represents the <b><i>amount</i></b> of time spent for something. This 
  * amount is defined as the <u><i>time difference</i></u> of the start to the 
  * end.
- * @author nicher
  */
 public class Duration {
 	
-	// -- generated attribute, constant + association declarations ----------
-	private double time;
+	private long time;
 	
 	private String unit;
 	
-	// -- generated constructors --------------------------------------------
-	
 	/**
 	 * Initializing constructor for class '<em><b>Duration</b></em>'.
-	 * @param	time	{@link double.getTime()}
+	 * @param	time	{@link long.getTime()}
 	 * @param	unit	{@link String.getUnit()}
 	 */
-	public Duration(double time, String unit) {
+	public Duration(long time, String unit) {
 		this.time = time;
 		this.unit = unit;
 	}
 	
-	// -- generated association + attribute accessors -----------------------
 	/**
 	 * Returns the value of attribute '<em><b>time</b></em>'
 	 */
-	public double getTime() {
+	public long getTime() {
 		return time;
 	}
 	
@@ -47,7 +41,6 @@ public class Duration {
 		return unit;
 	}
 	
-	// -- own code implementation -------------------------------------------
 	/* PROTECTED REGION ID(java.datatype.own.code.implementation._16_0_b6f02e1_1231923224937_24486_549) ENABLED START */
 	// TODO: put your own implementation code here
 	/* PROTECTED REGION END */
